+++ packages/solid/components.ts
@@ -1,2 +1,2 @@
-export { extend, getComponentCatalogue } from "./src/elements/index.js"
+export { extend, getComponentCatalogue } from "./src/elements/catalogue.js"
 export type { ExtendedComponentProps, ExtendedIntrinsicElements, OpenTUIComponents } from "./src/types/elements.js"
